organize & remove unused class
