+++ src/main/java/graphql/annotations/processor/exceptions/CannotCastMemberException.java
@@ -12,7 +12,7 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  */
-package graphql.annotations.processor.searchAlgorithms;
+package graphql.annotations.processor.exceptions;
 
 
 public class CannotCastMemberException extends RuntimeException {
